ai/requirements.txt
# ── Core ML ────────────────────────────────────────────────────
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
transformers==4.41.0
tokenizers==0.19.1
datasets==2.19.0
accelerate==0.30.0

# ── Classical ML ───────────────────────────────────────────────
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
scipy==1.13.0
numpy==1.26.4

# ── NLP utilities ──────────────────────────────────────────────
spacy==3.7.4
nltk==3.8.1
sentencepiece==0.2.0

# ── Audio processing ───────────────────────────────────────────
librosa==0.10.2
soundfile==0.12.1

# ── Video & image processing ───────────────────────────────────
opencv-python-headless==4.9.0.80
ffmpeg-python==0.2.0
mediapipe==0.10.14
pillow==10.3.0
imageio==2.34.1

# ── Model optimization ─────────────────────────────────────────
onnx==1.16.0
onnxruntime==1.18.0
onnxruntime-gpu==1.18.0

# ── Experiment tracking ────────────────────────────────────────
mlflow==2.13.0
wandb==0.17.0

# ── Data versioning ────────────────────────────────────────────
dvc==3.51.0
dvc-s3==3.2.0

# ── Calibration & evaluation ───────────────────────────────────
netcal==1.3.5

# ── Queue integration ──────────────────────────────────────────
celery==5.4.0
redis==5.0.4

# ── Utilities ──────────────────────────────────────────────────
python-dotenv==1.0.1
structlog==24.1.0
tqdm==4.66.4
